+++ src/main/java/com/trip/tripshorts/video/dto/PresignedUrlResponse.java
@@ -1,0 +1,11 @@
+package com.trip.tripshorts.video.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class PresignedUrlResponse {
+    private String presignedUrl;
+    private String filename;
+}
