+++ src/test/java/com/trip/tripshorts/TripShortsBeApplicationTests.java
@@ -1,6 +1,6 @@
 package com.trip.tripshorts;
 
-import com.trip.tripshorts.video.service.S3Service;
+import com.trip.tripshorts.util.S3Service;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
